rename xunit to suites (fixes #435, via #439)
