+++ allure-generator/src/main/java/io/qameta/allure/ConfigurationBuilder.java
@@ -23,11 +23,11 @@ import io.qameta.allure.mail.MailPlugin;
 import io.qameta.allure.owner.OwnerPlugin;
 import io.qameta.allure.retry.RetryPlugin;
 import io.qameta.allure.severity.SeverityPlugin;
+import io.qameta.allure.suites.SuitesPlugin;
 import io.qameta.allure.summary.SummaryPlugin;
 import io.qameta.allure.tags.TagsPlugin;
 import io.qameta.allure.timeline.TimelinePlugin;
 import io.qameta.allure.widget.WidgetsPlugin;
-import io.qameta.allure.xunit.XunitPlugin;
 
 import java.util.ArrayList;
 import java.util.Arrays;
@@ -63,7 +63,7 @@ public class ConfigurationBuilder {
                 new HistoryTrendPlugin(),
                 new GraphPlugin(),
                 new TimelinePlugin(),
-                new XunitPlugin(),
+                new SuitesPlugin(),
                 new ReportWebPlugin(),
                 new TestsResultsPlugin(),
                 new AttachmentsPlugin(),
